fixed issue with expired session

+++ routes/poll.js
@@ -19,22 +19,30 @@ pollRoutes.route("/api/polls")
 })
 //----- Create poll
 .post((req, res) => {
-  // Format options string
-  let formattedOptions = Format.formatOptions(req.body.options);
+  // Check for valid session
+  if(req.user) {
+    // Format options string
+    let formattedOptions = Format.formatOptions(req.body.options);
 
-  // Create new poll
-  let newPoll = new Poll({
-    userId: req.user._id,
-    topic: req.body.topic,
-    options: formattedOptions,
-  })
+    // Create new poll
+    let newPoll = new Poll({
+      userId: req.user._id,
+      topic: req.body.topic,
+      options: formattedOptions,
+    })
 
-  // Save new poll
-  newPoll.save()
-  .then(savedPoll => {
-    res.json({ success: true })
-  })
-  .catch(err => console.log(err));
+    // Save new poll
+    newPoll.save()
+    .then(savedPoll => {
+      res.json({ success: true })
+    })
+    .catch(err => console.log(err));
+  } else {
+    res.json({
+      success: false,
+      message: "Session has expired"
+    })
+  }
 });
 
 //----- Return all polls for given user
@@ -70,11 +78,19 @@ pollRoutes.route("/api/poll")
 })
 //----- Delete specified poll
 .delete((req, res) => {
-  Poll.findByIdAndDelete(req.body.id)
-  .then(deletedPoll => {
-    res.json({ success: true });
-  })
-  .catch(err => console.log(err));
+  // Check for valid session
+  if(req.user) {
+    Poll.findByIdAndDelete(req.body.id)
+    .then(deletedPoll => {
+      res.json({ success: true });
+    })
+    .catch(err => console.log(err));
+  } else {
+    res.json({
+      success: false,
+      message: "Session has expired"
+    })
+  }
 }); 
 
 //----- Update poll votes
@@ -111,34 +127,50 @@ pollRoutes.put("/api/poll/vote", (req, res) => {
 
 //----- Add new option for given poll
 pollRoutes.put("/api/poll/option", (req, res) => {
-  Poll.findByIdAndUpdate(req.body.id, {
-    $push: {
-      options: {
-        value: req.body.newOption,
-        votes: 0
+  // Check for valid session
+  if(req.user) {
+    Poll.findByIdAndUpdate(req.body.id, {
+      $push: {
+        options: {
+          value: req.body.newOption,
+          votes: 0
+        }
       }
-    }
-  }, {
-    new: true
-  })
-  .then(updatedPoll => {
-    res.json({ success: true })
-  })
-  .catch(err => console.log(err));
+    }, {
+      new: true
+    })
+    .then(updatedPoll => {
+      res.json({ success: true })
+    })
+    .catch(err => console.log(err));
+  } else {
+    res.json({
+      success: false,
+      message: "Session has expired"
+    })
+  }
 });
 
 //----- Change poll status
 pollRoutes.put("/api/poll/status", (req, res) => {
-  let updatedClosed = (req.body.status === "close" ? true : false);
-  Poll.findByIdAndUpdate(req.body.id, {
-    closed: updatedClosed
-  }, {
-    new: true
-  })
-  .then(updatedPoll => {
-    res.json({ success: true })
-  })
-  .catch(err => console.log(err));
+  // Check for valid session
+  if(req.user) {
+    let updatedClosed = (req.body.status === "close" ? true : false);
+    Poll.findByIdAndUpdate(req.body.id, {
+      closed: updatedClosed
+    }, {
+      new: true
+    })
+    .then(updatedPoll => {
+      res.json({ success: true })
+    })
+    .catch(err => console.log(err));
+  } else {
+    res.json({
+      success: false,
+      message: "Session has expired"
+    })
+  }
 });
 
 module.exports = pollRoutes;
